redfish-core: Add multi host system support for redfish
The redfish core library api currently only supports platforms with
a single host/computer system. Multiple hosts of a platform cannot
be discovered and computer system specific redfish queries are not
supported with the current implementation.

This change adds initial support to discover and populate multi host
systems and support system info queries. This implementation is based on
the work of Naveen Moses and the discussion that has been started
on that review:
https://gerrit.openbmc.org/c/openbmc/bmcweb/+/52075

Furthermore, this patch is only one of multiple ones.
As of now, the idea is to change every header file separately,
in order to keep following reviews as small as possible.

Design:
System discovery via the inventory seems to be problematic, as has been
already discussed in the review linked. In short, this is due to
platforms adding systems in a non standard way to the inventory,
which makes discovery more complicated via the inventory.
Thus, the idea to use the ObjectMapper state objects seem to fit better,
as they should stay the same across platforms.
The dbus interface for the host state also helps figuring out,
whether we are on a single or on a multi host platform.

This implementation ignores the
'experimental-redfish-multi-computer-system' feature flag
entirely, since it reached EOL as stated in meson_options.txt.
The goal here is to integrate multi host support seamlessly into the
existing redfish-core library, without breaking single host and keeping
behavior on single host as is.

The redfish system uri name for a single host system is set by the user
via meson option and defaults to "system", without an index:

i.e : /redfish/v1/Systems/system

Multi host systems are identified via indices. The provided
REDFISH_SYSTEM_URI_NAME is used as base string, the index is appended to
it:

/redfish/v1/Systems/redfish_system_uri_name1
/redfish/v1/Systems/redfish_system_uri_name2
.
.
/redfish/v1/Systems/redfish_system_uri_nameN

Testing:
In order to test the implementation, the steps advised in TESTING.md
were followed.
Single host testing was performed on hardware and the romulus image
in qemu, multi host testing could only be performed in a
qemu environment (yosemite4), as of now.
The RedfishServiceValidator has not reported any new problems introduced
to single host, for yv4 validation succeeded.
POST and PATCH requests were tested on hardware via WebUI and worked as
expected.

Multi host test results on yv4 with the default REDFISH_SYSTEM_URI_NAME:

$ curl -k 'https://'"${BMC}"':'"${BMC_WEBPORT}"'/redfish/v1/Systems' \
 -H 'X-Auth-Token: '"$BMCWEB_SESSION_TOKEN"''

{
  "@odata.id": "/redfish/v1/Systems",
  "@odata.type": "#ComputerSystemCollection.ComputerSystemCollection",
  "Members": [
    {
      "@odata.id": "/redfish/v1/Systems/system1"
    },
    {
      "@odata.id": "/redfish/v1/Systems/system2"
    },
    {
      "@odata.id": "/redfish/v1/Systems/system3"
    },
    {
      "@odata.id": "/redfish/v1/Systems/system4"
    },
    {
      "@odata.id": "/redfish/v1/Systems/system5"
    },
    {
      "@odata.id": "/redfish/v1/Systems/system6"
    },
    {
      "@odata.id": "/redfish/v1/Systems/system7"
    },
    {
      "@odata.id": "/redfish/v1/Systems/system8"
    }
  ],
  "Members@odata.count": 8,
  "Name": "Computer System Collection"
}

$ curl -k \
'https://'"${BMC}"':'"${BMC_WEBPORT}"'/redfish/v1/Systems/system1' \
-H 'X-Auth-Token: '"$BMCWEB_SESSION_TOKEN"''

{
  "@odata.id": "/redfish/v1/Systems/system1",
  "@odata.type": "#ComputerSystem.v1_22_0.ComputerSystem",
  "Actions": {
    "#ComputerSystem.Reset": {
  "@Redfish.ActionInfo": "/redfish/v1/Systems/system1/ResetActionInfo",
  "target": "/redfish/v1/Systems/system1/Actions/ComputerSystem.Reset"
    }
  },
  "Bios": {
    "@odata.id": "/redfish/v1/Systems/system1/Bios"
  },
  "Boot": {
    "AutomaticRetryAttempts": 3,
    "AutomaticRetryConfig": "Disabled",
    "AutomaticRetryConfig@Redfish.AllowableValues": [
      "Disabled",
      "RetryAttempts"
    ],
    "BootSourceOverrideEnabled": "Disabled",
    "BootSourceOverrideMode": "UEFI",
    "BootSourceOverrideMode@Redfish.AllowableValues": [
      "Legacy",
      "UEFI"
    ],
    "BootSourceOverrideTarget": "None",
    "BootSourceOverrideTarget@Redfish.AllowableValues": [
      "None",
      "Pxe",
      "Hdd",
      "Cd",
      "Diags",
      "BiosSetup",
      "Usb"
    ],
    "RemainingAutomaticRetryAttempts": 3,
    "StopBootOnFault": "Never",
    "TrustedModuleRequiredToBoot": "Disabled"
  },
  "Description": "Computer System",
  "FabricAdapters": {
    "@odata.id": "/redfish/v1/Systems/system1/FabricAdapters"
  },
  "GraphicalConsole": {
    "ConnectTypesSupported": [
      "KVMIP"
    ],
    "MaxConcurrentSessions": 4,
    "ServiceEnabled": true
  },
  "Id": "system1",
  "Links": {
    "ManagedBy": [
      {
        "@odata.id": "/redfish/v1/Managers/bmc"
      }
    ]
  },
  "LogServices": {
    "@odata.id": "/redfish/v1/Systems/system1/LogServices"
  },
  "Memory": {
    "@odata.id": "/redfish/v1/Systems/system1/Memory"
  },
  "MemorySummary": {
    "TotalSystemMemoryGiB": 0.0
  },
  "Name": "system1",
  "PCIeDevices": [],
  "PCIeDevices@odata.count": 0,
  "PowerRestorePolicy": "AlwaysOn",
  "ProcessorSummary": {
    "Count": 0
  },
  "Processors": {
    "@odata.id": "/redfish/v1/Systems/system1/Processors"
  },
  "SerialConsole": {
    "IPMI": {
      "ServiceEnabled": true
    },
    "MaxConcurrentSessions": 15,
    "SSH": {
      "HotKeySequenceDisplay": "Press ~. to exit console",
      "Port": 2200,
      "ServiceEnabled": true
    }
  },
  "Status": {
    "Health": "OK",
    "State": "Enabled"
  },
  "Storage": {
    "@odata.id": "/redfish/v1/Systems/system1/Storage"
  },
  "SystemType": "Physical"
}

Change-Id: I82d59487b7c17b22cd638acd8f687f31c96ca156
Signed-off-by: Oliver Brewka <mox669.dev@gmail.com>