New method to retrieve DBUS resources for systems
In an effort to support multi-host in bmcweb, a new method to retrieve
dbus resources is needed. Since the long-term goal for bmcweb is to get
rid of hardcoded URIs as much as possible to align with the redfish
spec, this method shall handle single host as well.

Initially bmcweb hardcoded object paths and services to retrieve various
resources via dbus method calls. However, the introduction of multi host
platforms also introduces new service names and object paths.
Hardcoding them does not cut it anymore, since we don't necessarily
know how they are called, nor should we do any assumptions on naming
schemes by default.

This method follows this idea and lets DBUS return needed paths and
services, if any, primarily based on a systemName the user sends
a request for.

Change-Id: I67c17c3dd7a354fa9a2ebbc56d4def7a7e788909
Signed-off-by: Oliver Brewka <mox669.dev@gmail.com>