Add support for multi-host in systems.hpp
Add support for multi-host GET, POST and PATCH
request handling under the
/redfish/v1/Systems/{computerSystemId}/ redfish resource.

The change implements the getComputerSystemIndex which leverages
the xyz.openbmc_project.ManagedHost interface to retrieve the
index corresponding to the host the request was made for, when
on a multi-host machine. On single-host the index always defaults
to 0. It is not the perfect solution, but it is a starting point
to introduce multi-host support to bmcweb, allowing for power control,
via redfish and in general something we can build from.
More efficient ways of doing things are already discussed,
but will need more time to design properly.

The implementation relies on the
experimental-redfish-multi-computer-system meson flag to give bmcweb an
indication, whether it is running on single- or multi-host and to drop
unsupported redfish resources for the time being.

All multi-host supported redfish URIs can be found in this listing:
https://gerrit.openbmc.org/c/openbmc/bmcweb/+/76118

Tested:
GET route has been tested with redfish service validator in qemu for
the romulus and yosemite4 machine. Validator passed, no regression
observed.

POST route has been manually tested in web-ui on single-host hardware.
All possible power controls worked as expected. On multi-host the
expected changes have been observed on dbus in qemu.

PATCH route has been updated via curl
manually both on multi- and single-host for every property, that this
patch implements. After each update the
expected change has been observed on dbus.

Change-Id: I67c17c3dd7a354fa9a2ebbc56d4def7a7e788909
Signed-off-by: Oliver Brewka <mox669.dev@gmail.com>