Add multi-host support in systems.hpp
This change introduces multi-host support to systems.hpp.
Previously no request handler in systems.hpp was capable of handling
the request for a multi-host system (this was indicated by the
BMCWEB_EXPERIMENTAL_REDFISH_MULTI_COMPUTER_SYSTEM constant inside the
handler used to return a 404 resource not found). This change fixes all
occurances in systems.hpp and enables bmcweb to handle a request for
multi-host systems.

Testing: todo

Change-Id: I67c17c3dd7a354fa9a2ebbc56d4def7a7e788909
Signed-off-by: Oliver Brewka <mox669.dev@gmail.com>