update service: related item by association
Use the mapper to find the associated inventory item for a software
version via 'running' / 'ran_on' association [1].

Then link to it via 'RelatedItem' property.

The problem here is to find the redfish URL of the inventory item, if
such an URL exists.

In this patch the implementation is first done for baseboards which are
managed hosts.

References:
[1] https://github.com/openbmc/docs/blob/master/designs/code-update.md

Tested: Querying the fw inventory item shows a related item for the
correct system.

```
curl https://${bmc}/redfish/v1/UpdateService/FirmwareInventory/HostSPIFlash_9701
```

```
{
  "@odata.id": "/redfish/v1/UpdateService/FirmwareInventory/HostSPIFlash_9701",
  "@odata.type": "#SoftwareInventory.v1_1_0.SoftwareInventory",
  "Description": "Unknown image",
  "Id": "HostSPIFlash_9701",
  "Name": "Software Inventory",
  "RelatedItem": [
    {
      "@odata.id": "/redfish/v1/Systems/Tyan_S8030_Baseboard"
    }
  ],
  "RelatedItem@odata.count": 1,
  "Status": {
    "Health": "Warning",
    "HealthRollup": "OK",
    "State": "Disabled"
  },
  "Updateable": true,
  "Version": "Unknown"
}
```

Change-Id: I2057ef57c1d2fd9fc8a0d6951773a317632afc8d
Signed-off-by: Alexander Hansen <alexander.hansen@9elements.com>